refactor: retira do service e controller do gerente responsabilidades que não lhe pertencem
