+++ src/gerente/gerente.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { GerenteService } from './gerente.service';
 import { GerenteController } from './gerente.controller';
-import { ContaBancariaFactory } from 'src/contas/contaBancaria.factory';
+import { ContaBancariaFactory } from 'src/contaBancaria/contaBancaria.factory';
 
 @Module({
   providers: [GerenteService, ContaBancariaFactory],
